#Conflicts:
#	projetoFinal/contentSecView.swift
#	projetoFinal/view/jogoVIew/jogoCardView.swift

+++ projetoFinal/contentSecView.swift
@@ -2,21 +2,90 @@ import SwiftUI
 import SwiftData
 
 struct ContentSecView: View {
+    
     @Environment(\.modelContext) private var modelContext
     @Query(sort: \Jogo.data) private var jogos: [Jogo]
-    
+    @Environment(\.dismiss) private var dismiss
+    @State private var jogoSelecionado: Jogo? = nil
+    @State private var mostrarEdicao = false
+    @State private var mostrarConfirmacaoDeletar = false
+    @State private var criandoJogo: Bool = false
     var body: some View {
         NavigationStack {
-            List {
-                ForEach(jogos) { jogo in
-                    NavigationLink(destination: JogoEditingView(jogo: jogo)
-                        .environment(\.modelContext, modelContext)) {
-                        JogoCardView(jogo: jogo)
+            VStack {
+                    NavigationStack {
+                        VStack {
+                            ScrollView {
+                                LazyVStack(spacing: 10) {   // espaço entre os cards
+                                    ForEach(jogos) { jogo in
+                                        JogoCardView(jogo: jogo)
+                                            .onTapGesture {
+                                                withAnimation {
+                                                    jogoSelecionado = jogo
+                                                }
+                                            }
+                                    }
+                                }
+                                .padding() // margem geral
+                            }
+                        }
+                        .navigationTitle("Admin - Jogos")
+                        .toolbar {
+                            ToolbarItem(placement: .navigationBarTrailing) {
+                                Button {
+                                    jogoSelecionado = nil
+                                    criandoJogo = true
+                                } label: {
+                                    Image(systemName: "plus")
+                                }
+                            }
+                        }
                     }
                 }
-                .onDelete { indexSet in
-                    for index in indexSet {
-                        let jogo = jogos[index]
+                
+              
+                if let jogo = jogoSelecionado {
+                    Divider()
+                        .padding(.vertical, 8)
+                       
+
+                    VStack(spacing: 12) {
+                        JogoCardView(jogo: jogo)
+                            .opacity(0.5)
+                            .padding(.horizontal)
+
+                        HStack(spacing: 30) {
+                            
+                            Button("Editar") {
+                                mostrarEdicao = true
+                            }
+                            .buttonStyle(.borderedProminent)
+
+                            Button("Excluir") {
+                                mostrarConfirmacaoDeletar = true
+                            }
+                            .buttonStyle(.bordered)
+                            .tint(.red)
+                        }
+                    }
+                    .padding()
+                    .background(.regularMaterial)
+                    .cornerRadius(20)
+                    .padding(.horizontal)
+                }
+            }
+            .sheet(isPresented: $criandoJogo){
+                SecFormView()
+                    .environment(\.modelContext, modelContext)
+            }
+            .sheet(isPresented: $mostrarEdicao) {
+                JogoEditingView()
+                    .environment(\.modelContext, modelContext)
+            }
+            
+            .alert("Deseja excluir este jogo?", isPresented: $mostrarConfirmacaoDeletar) {
+                Button("Excluir", role: .destructive) {
+                    if let jogo = jogoSelecionado {
                         modelContext.delete(jogo)
                     }
                 }
